Add files via upload
This project is an end-to-end data pipeline leveraging Azure services such as Azure Data Factory, Azure Databricks, and Azure Storage. Additionally, Power BI was integrated into the pipeline. This project addressed underutilized sales data in the pizza industry, with the goal of optimizing operations and increasing profitability through data-driven decision-making.